+++ test/unit/adapter/test_adapt_verification_rules.py
@@ -20,19 +20,6 @@ def get_valid_pipeline():
     return opt_graph, pipeline, adapter
 
 
-@pytest.mark.parametrize('rule', DEFAULT_DAG_RULES)
-def test_adapt_verification_rules_dag(rule):
-    """Test that dag verification rules behave as expected with new adapter.
-    They accept any graphs, so the new adapter must see them as native
-    and shouldn't change them on the call to adapt."""
-
-    opt_graph, pipeline, adapter = get_valid_pipeline()
-    adapted_rule = adapter.adapt_func(rule)
-
-    assert adapted_rule(opt_graph)
-    assert id(rule) == id(adapted_rule)
-
-
 @pytest.mark.parametrize('rule', SOME_PIPELINE_RULES)
 def test_adapt_verification_rules_pipeline(rule):
     """Test that pipeline verification rules behave as expected with new adapter."""
